requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Monitoring
psutil>=5.9.0

# Utilities
numpy>=1.24.0
requests>=2.31.0

# Nouvelles dépendances V3
timm>=0.9.0
albumentations>=1.3.0

# Support vidéo
opencv-python>=4.8.0
facenet-pytorch>=2.5.2

